backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10

# Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & WebSocket
httpx==0.27.2
websockets==13.1
python-multipart==0.0.12
sse-starlette==2.1.3

# Auth (for future)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Parsing & Processing
pyyaml==6.0.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Code Quality
black==24.10.0
flake8==7.1.1

